Use relative paths for node-pre-gyp etc. in npm install scripts
And. Uh. Run npm install node-pre-gyp in the post-install script.
This is ridiculous; the install script should only run after all the
dependencies have been installed. But for some reason, without that
line, at the time it was running, node-pre-gyp had not been installed.
With that line it works. I don't have time right now to figure out
why.